Remove kncloudevents.CloudEventsRequest
